[1.2.0] Split `InlineRewriter` in submodules
